html/semantics/forms/the-select-element/select-setcustomvalidity.html

<!DOCTYPE HTML>
<title>select setCustomValidity</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<select id='test'></select>

<script>

test(() => {
  let select = document.getElementById("test");
  assert_false(select.validity.customError);
  select.setCustomValidity("custom error");
  assert_true(select.validity.customError);
}, "select setCustomValidity is correct")

</script>
